requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==2.1.1
click==8.1.8
distro==1.9.0
elevenlabs==1.50.3
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
websockets==14.1
Werkzeug==3.1.3
yarl==1.18.3
